Add manipulation approval interface to rviz
This is currently implemented as part of the nlu panel, but it should probably be separated into its own panel.

The approval request subscriber is currently expecting a string message, but this should be updated to the custom ManipulationApprovalRequest message that includes the robot_id (for response publisher namespacing) and the image.